requirements.txt
Python==3.13.2
Flask==3.1.0
PyMuPDF==1.25.0
google-genai
python-dotenv
